+++ src/colours/taiga-dark.ts
@@ -1,0 +1,86 @@
+const darkColours = {
+	blue: {
+		0: "#1762B6",
+		1: "#3A81D7",
+		2: "#59A0F9",
+		3: "#91C1FF",
+		4: "#C9E0FF",
+	},
+	green: {
+		0: "#007742",
+		1: "#2C965D",
+		2: "#51B67A",
+		3: "#72D699",
+		4: "#93F8B8",
+	},
+	purple: {
+		0: "#773CBF",
+		1: "#945CE1",
+		2: "#B17EFF",
+		3: "#C9ACFF",
+		4: "#E3D6FF",
+	},
+	red: {
+		0: "#BA0D01",
+		1: "#DE3E2D",
+		2: "#FF6551",
+		3: "#FFA191",
+		4: "#FFD2CA",
+	},
+	teal: {
+		0: "#00736E",
+		1: "#14938D",
+		2: "#43B2AC",
+		3: "#67D2CC",
+		4: "#89F4ED",
+	},
+	yellow: {
+		0: "#825B00",
+		1: "#A77600",
+		2: "#CD9200",
+		3: "#F0B135",
+		4: "#FFD896",
+	},
+};
+
+export const taiga = {
+	name: "Taiga",
+	type: "dark",
+	ui: {
+		0: "#000924",
+		1: "#10213e",
+		2: "#283b5a",
+		3: "#425677",
+		highlight: "#00681b",
+		shadow: "#0F171F",
+	},
+	text: {
+		1: "#D3DEF5",
+		2: "#A4AEC4",
+		3: "#5A6377",
+		4: "#3f475a",
+		highlight: "#72D699",
+	},
+	...darkColours,
+};
+
+export const taigaTrees = {
+	name: "Taiga Trees",
+	type: "dark",
+	ui: {
+		0: "#001100",
+		1: "#112810",
+		2: "#2a4228",
+		3: "#374f35",
+		highlight: "#00681b",
+		shadow: "#0F171F",
+	},
+	text: {
+		1: "#D4E3D0",
+		2: "#b4c3b1",
+		3: "#778573",
+		4: "#5a6758",
+		highlight: "#72D699",
+	},
+	...darkColours,
+};